Improve: Stub 미구현 메소드 구현
